+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 
+### Other Changes
+
+- Added a more extensive demo app.
+
 ## [1.0.2](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v1.0.2) - 2022-06-21
 
 ### New features